+++ ais/ais_test.go
@@ -3,6 +3,7 @@ package ais_test
 import (
 	"bytes"
 	"encoding/json"
+	"fmt"
 	"github.com/ilder-as/go-barentswatch-ais/ais"
 	"golang.org/x/oauth2"
 	"io"
@@ -14,6 +15,27 @@ import (
 	"time"
 )
 
+func oauthSpoofMW(handlerFunc http.HandlerFunc) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		if strings.Contains(r.URL.Path, "token") {
+			w.Header().Add("Content-Type", "application/json")
+			token, _ := json.Marshal(oauth2.Token{AccessToken: "x", TokenType: "Bearer", Expiry: time.Now().Add(time.Hour)})
+			w.Write(token)
+			return
+		}
+		handlerFunc(w, r)
+	}
+}
+
+func server(t *testing.T, handlerFunc http.HandlerFunc) *httptest.Server {
+	sv := httptest.NewUnstartedServer(handlerFunc)
+	sv.EnableHTTP2 = true
+	sv.Start()
+
+	return sv
+}
+
+// fixtureServer serves the content of the file as the response payload, Content-Type application/json
 func fixtureServer(t *testing.T, filename string) *httptest.Server {
 	f, err := os.Open(filename)
 	if err != nil {
@@ -39,6 +61,118 @@ func fixtureServer(t *testing.T, filename string) *httptest.Server {
 	sv.Start()
 
 	return sv
+}
+
+// hijackServer serves the content of the provided file byte for byte as the HTTP response. It must therefore
+// include all headers etc.
+func hijackServer(t *testing.T, filename string) *httptest.Server {
+	f, err := os.Open(filename)
+	if err != nil {
+		t.Fatalf("unable to load test fixture \"%s\": %s", filename, err)
+	}
+
+	buf := bytes.Buffer{}
+	if _, err = io.Copy(&buf, f); err != nil {
+		t.Fatalf("unable to copy data from file: %s", err)
+	}
+	f.Close()
+
+	sv := httptest.NewUnstartedServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		if strings.Contains(r.URL.Path, "token") {
+			w.Header().Add("Content-Type", "application/json")
+			token, _ := json.Marshal(oauth2.Token{AccessToken: "x", TokenType: "Bearer", Expiry: time.Now().Add(time.Hour)})
+			w.Write(token)
+			return
+		}
+		c, _, err := http.NewResponseController(w).Hijack()
+		if err != nil {
+			t.Fatalf("error in response controller: %s", err)
+		}
+		io.Copy(c, &buf)
+		c.Close()
+	}))
+	sv.EnableHTTP2 = true
+	sv.Start()
+
+	return sv
+}
+
+func Test_GetAis(t *testing.T) {
+	filename := "testdata/get_ais.txt"
+	sv := fixtureServer(t, filename)
+	defer sv.Close()
+
+	urls := ais.DefaultURLs()
+	urls.OAuthBase = sv.URL
+	urls.APIBase = sv.URL
+
+	client := ais.NewClient("", "", urls)
+	_ = client
+
+	stream, err := client.GetAis()
+	if err != nil {
+		t.Fatal(err)
+	}
+	ch, err := stream.UnmarshalStream()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	numResults := 0
+	for a := range ch {
+		if a.IsZero() {
+			t.Error("found zero record")
+		}
+		numResults++
+	}
+
+	if numResults <= 0 {
+		t.Errorf("expected to find results, found %d", numResults)
+	}
+
+	if err = stream.Error(); !ais.IsEOF(err) {
+		t.Fatalf("expected EOF, got \"%s\"", err)
+	}
+}
+
+func Test_PostAis_NoneIncluded(t *testing.T) {
+	filename := "testdata/post_ais_none_included_all.txt"
+	sv := hijackServer(t, filename)
+	defer sv.Close()
+
+	urls := ais.DefaultURLs()
+	urls.OAuthBase = sv.URL
+	urls.APIBase = sv.URL
+
+	client := ais.NewClient("", "", urls)
+	_ = client
+
+	stream, err := client.PostAis(ais.FilterInput{})
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	fmt.Println(stream.Response.Status)
+
+	t.SkipNow()
+
+	ch, err := stream.UnmarshalStream()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	numResults := 0
+	for range ch {
+		numResults++
+	}
+
+	if numResults > 0 {
+		t.Errorf("expected to find results, found %d", numResults)
+	}
+
+	if err = stream.Error(); err != nil {
+		t.Fatalf("expected error, %s", err)
+	}
 }
 
 func Test_PostCombined_FullGeojson(t *testing.T) {
@@ -209,5 +343,69 @@ func Test_PostCombined_Broken(t *testing.T) {
 
 	if err = stream.Error(); err == nil || ais.IsEOF(err) {
 		t.Fatalf("expected non-EOF error, got \"%s\"", err)
+	}
+}
+
+func Test_GetSSEAis(t *testing.T) {
+	filename := "testdata/get_sse_ais.txt"
+	sv := fixtureServer(t, filename)
+	defer sv.Close()
+
+	urls := ais.DefaultURLs()
+	urls.OAuthBase = sv.URL
+	urls.APIBase = sv.URL
+
+	client := ais.NewClient("", "", urls)
+
+	stream, err := client.GetSSEAis()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	ch, err := stream.UnmarshalStream()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	num := 0
+	for data := range ch {
+		if data.IsZero() {
+			t.Error("got zero data")
+		}
+		num++
+	}
+
+	if num <= 0 {
+		t.Errorf("expected > 0 results, got %d", num)
+	}
+
+	if err = stream.Error(); !ais.IsEOF(err) {
+		t.Fatalf("expected EOF, got \"%s\"", err)
+	}
+}
+
+func TestClient_GetLatestAis(t *testing.T) {
+	filename := "testdata/get_latest_ais.txt"
+	sv := hijackServer(t, filename)
+	defer sv.Close()
+
+	urls := ais.DefaultURLs()
+	urls.OAuthBase = sv.URL
+	urls.APIBase = sv.URL
+
+	client := ais.NewClient("", "", urls)
+
+	res, err := client.GetLatestAis(ais.Since(time.Now()))
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	data, err := res.Unmarshal()
+	if err != nil {
+		t.Fatalf("error unmarshalling: %s", err)
+	}
+
+	if len(data) <= 0 {
+		t.Errorf("expected response to be non-empty list, list had %d members", len(data))
 	}
 }